Opt-out fwaas, removed lbaas, dhcp6-pd removed
The fwaas-feature is now opt-out-able. The hiera-key
`ntnuopenstack::neutron::fwaas::enabled` can be set to false to disable fwaas.

The neutron-lbaas is deprecated for a while, and is removed in Train. It is thus
removed entirly here.

The cleaning of the dhcpv6-pd services used earlier, and left to clean in Stein
is now removed. This is the last step of phasing out dhcpv6-pd.

BGP DR-Agents are now removed if they dont get a Router-ID.

Old-style config for ntnuopenstack::neutron::external::connections is also
removed. All our platform uses new-style, so this should be fine.